requirements.txt
asgiref==3.6.0
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Django==4.2
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.3.2
geos==0.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.25.2
pandas==2.0.3
Pillow==9.5.0
Pygments==2.15.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
views==0.3
Werkzeug==2.3.6
